requirements.txt
gunicorn
uvicorn
supabase
fastapi
openai
google-generativeai
pandas
streamlit
python-dotenv